Added mbfl_buffer_converter_dtor() Added more check for header files
git-svn-id: http://svn.sourceforge.jp/svnroot/php-i18n/libmbfl/trunk@134 76bb4867-1e22-0410-9289-e9e6e9b3b6f2

+++ mbfl/mbfl_allocators.h
@@ -31,15 +31,13 @@
 #ifndef MBFL_ALLOCATORS_H
 #define MBFL_ALLOCATORS_H
 
-#include <stddef.h>
-
 typedef struct _mbfl_allocators {
-	void *(*malloc)(size_t);
-	void *(*realloc)(void *, size_t);
-	void *(*calloc)(unsigned int, size_t);
+	void *(*malloc)(unsigned int);
+	void *(*realloc)(void *, unsigned int);
+	void *(*calloc)(unsigned int, unsigned int);
 	void (*free)(void *);
-	void *(*pmalloc)(size_t); 
-	void *(*prealloc)(void *, size_t);
+	void *(*pmalloc)(unsigned int); 
+	void *(*prealloc)(void *, unsigned int);
 	void (*pfree)(void *);
 } mbfl_allocators;
